Create honeycrisp modularized js
