+++ app/assets/javascripts/honeycrisp.module.js
@@ -1,0 +1,5 @@
+var Honeycrisp = require('./honeycrisp');
+
+module.exports = Honeycrisp
+
+
